agents/pyproject.toml
[project]
name = "atlas-agents"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
